+++ modules/main/src/main/java/com/nisovin/shopkeepers/compat/Compat.java
@@ -44,6 +44,7 @@ public final class Compat {
 	// is paper-specific.
 	static {
 		// Registered in the order from latest to oldest.
+		register(new CompatVersion("1_21_R6", "1.21.6", "164f8e872cb3dff744982fca079642b2"));
 		register(new CompatVersion("1_21_R5_paper", "1.21.5", "7ecad754373a5fbc43d381d7450c53a5"));
 		register(new CompatVersion("1_21_R5", "1.21.5", "7ecad754373a5fbc43d381d7450c53a5"));
 		register(new CompatVersion("1_21_R4", "1.21.4", "60ac387ca8007aa018e6aeb394a6988c"));