+++ modules/v1_21_R5_paper/src/main/java/com/nisovin/shopkeepers/compat/v1_21_R5_paper/CompatProviderImpl.java
@@ -349,7 +349,7 @@ public final class CompatProviderImpl implements CompatProvider {
 		var nmsItem = net.minecraft.world.item.ItemStack.parse(
 				CraftRegistry.getMinecraftRegistry(),
 				convertedItemTag
-		).orElse(net.minecraft.world.item.ItemStack.EMPTY);
+		).orElseThrow();
 		return CraftItemStack.asCraftMirror(nmsItem);
 	}